+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Restrict supported compression algorithms to `gzip`
+
 ## HCloud v2.1.0 (2024-07-28)
 
 - Add `included_traffic` attribute to `ServerType#prices` and `LoadBalancerType#prices`